Make project runnable on clone
By default VSIX projects store the settings necessary to run and debug
the project in the .csproj.user file.  This is excluded by default from
most source code control systems.  Hence it produces the nasty effect of
works on your machine but fails whenever anyone clones the project.
This change just moves the appropriate settings into the .csproj file so
that F5 works on clone