Fixes encouragement at the end of the document
When the caret is at the very end of the document the trigger point
comes back as the end point of the document.  The equivalent of

```
new Snapshot(snapshot, snapshot.Length);
```

We then created a tracking span of trigger point + 1.  This creates a
point past the end of the buffer and hence caused an exception to be
raised.  The fix is to simply have a 0 length tracking span here.