Remove javascript signatures from encouragement
The Javascript IDE implementation has a bug where by it will insert its
signature help into any ISignatureHelpSession that is started.  This
includes the session that is started by the Enouragement extension.
This causes encouragement tips to be intermixed with javascript
signatures

Due to the implementation of the Javascript provider the only suitable
short term fix is to a) run after them and b) just remove their
signatures from our session.

This fixes #16

As a consequence it also fixes #1 as well.